+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMedia.kt
@@ -9,4 +9,5 @@ fun String.toInputMediaFileAttachmentName() = "attach://$this"
 interface InputMedia {
     val type: String
     val file: InputFile
+    val media: String
 }