+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMediaAnimation.kt
@@ -22,8 +22,6 @@ data class InputMediaAnimation(
     override val type: String = "animation"
 
     @SerialName(mediaField)
-    val media: String = when (file) {
-        is FileId -> file.fileId
-        is MultipartFile -> file.fileId.toInputMediaFileAttachmentName()
-    }
+    override val media: String
+    init { media = file.fileId } // crutch until js compiling will be fixed
 }