+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMediaDocument.kt
@@ -19,8 +19,6 @@ data class InputMediaDocument(
     override val type: String = "document"
 
     @SerialName(mediaField)
-    val media: String = when (file) {
-        is FileId -> file.fileId
-        is MultipartFile -> file.fileId.toInputMediaFileAttachmentName()
-    }
+    override val media: String
+    init { media = file.fileId } // crutch until js compiling will be fixed
 }